Simplified pet module and event processing
Removes the need for a separate module to track the event and cuts 4 events down to 1. Every individual event has edge cases where scanning the entire pet journal is needed, so instead of programming a fix for each edge case, just look for pet journal events instead. While they fire on every filter change, that's easily worked around by looking at the total number of pets the player has. If the number is the same as the previous check, we can skip further processing.