Added DB post-processing
Initial concept is for objects to have functions applied and checked in their collected state, but only if the object was loaded in the first place.

The entry-point here is not very good and should eventually be reworked such that DB processing and post-processing is done asynchronously after login to reduce impact to login times.