Added spell requirement and select list refreshes
This was added primarily to support zone time phasing with Zidormi. For example, the blasted lands have some quests that require you to talk to Zidormi to return to the past. She applies an aura to the player that persists until the player leaves the zone or talks to her again to change the timeline.